re-structure python module for publishing
