+++ .gitignore
@@ -70,7 +70,7 @@ typings/
 # build files and other autogenerated files
 */build/
 */bindings/
-*/src/
+tree-sitter-usfm/src/
 */binding.gyp
 */Cargo.toml
 */*.wasm
