--- python-usfm-parser/grammar_rebuild.py
@@ -1,15 +1,0 @@
-''' This code is to be run only if we update the grammar and need to use the new one
- and need to be used only in dev(not in the published library)'''
-
-from tree_sitter import Language
-
-Language.build_library(
-  # Store the library in the `build` directory
-  'build/my-languages.so',
-
-  # Include one or more languages
-  [
-    # '../tree-sitter-exp',
-    '../tree-sitter-usfm'
-  ]
-)
