+++ python-usfm-parser/src/usfm_grammar.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+usfm-grammar = usfm_grammar.__main__:main
