+++ python-usfm-parser/src/usfm_grammar.egg-info/requires.txt
@@ -1,0 +1,7 @@
+tree-sitter==0.20.0
+lxml==4.9.1
+
+[dev]
+rnc2rng
+jupyterlab
+pytest
