+++ python-usfm-parser/src/usfm_grammar.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+grammar_rebuild
+usfm_grammar
